NIFI-9686: Renamed SNMP integration tests correctly
